reorganize benchmarks for flamegraph analysis
